[369897] Improved support for rename refactoring Need specialized support for IDependentElementsCalculator.
I added specialized toString support for the GenModel classes so that they
print the unresolved reference to the underlying Ecore model; this makes it
easier to identify what instance it represents while debugger.

I improved the Convert to Xcore action to not import implicitly aliased data
types and not to define annotation directives for those things visible in
xcore.lang.  Also needed to avoid putting wrapper GenModels for Ecore,
XMLType,
and XMLNamespace in the xcore resource itself.

Need to normalize instance type names so they validate correctly.