[474574] Support validation of paths in the GenModel
Support live validation to show the constraint violations in the tree
view and in the properties view.  Show diagnostic decorators in the
Xcore editor's properties view. Use a command when reconciling the
GenModel so that the editor is dirty when the model has been reconciled.
Improve placement of issues reported in the Xcore editor for values
specified by annotations.  The build is disabled when the GenModel has
validation errors.