[438123] Provide for extensible cell editor support
GenDataType and GenFeature both support propertyEditorFactory.   These
properties are implemented to read directly from GenModel annotations,
in which case they don't serialize into the *.genmodel.  The
GenFeature.propertyEditorFactory, when not specifically set or present
as an annotation, derives its value from the value of this property on
the GenDataType.  Validation is supported.  The value must correspond to
a registered IPropertyEditorFactory.   An IPropertyEditorFactory
supports creating cell editors, but it's a generic API.  The two primary
IItemPropertyDescriptor implementation classes implement
IPropertyEditorFactory.Provider and have new constructors for the
additional argument; the templates generate the right call when the
property is present.  The provider can return a URI for a registered
factory (there is of course a new extension point to register them) or
can return a factory instance.  PropertyDescriptor tests for this
interface and delegates to the item property descriptor to get the
factory; if one is provided, the factory is used.  If that actually
creates a cell editor, that is of course used. Otherwise the default
behavior kicks in to create the cell editor. IPropertyEditorFactory also
support specializing the labels used for the values, and supports
validation that the URI is well formed.  Factory lookup is done by
trimming the query, fragment, and successive segments of the URI, so the
URI can be used to pass additional arguments to the factory.  Reflective
item providers look up the annotations so they also support specialized
property editors.  IPropertyEditorFactory cannot be implemented
directly.  One of the provided classes must be extended.  The
implementation classes support suggesting choices of URI values
appropriate for the context, or even being omitted because the context
doesn't support that type of cell editor.

To test out these concepts and the design, a DateConversionDelegate is
implemented.  It goes the whole nine yards to support annotation
validation and specialized property descriptors to suggest formatting
choices.  It basically supports specialized serialization conversion for
long, Long, Date, java.sql.Date,  Calendar, GregorianCalendar, and
XMLGregorianCalendar.   With properly serialized Date types available,
NebulaDatePropertyEditorFactory is implemented to demonstrate the
concepts.  It supports using a Nebula CDateTime cell editor on features
or data types with an instance type convertible to a Date, i.e., the
ones listed previously as also supported by the date conversion
delegate.  This implementation used reflection to avoid a hard
dependency on Nebula; it's an optional greedy dependency.  The
implementation illustrates how the propertyEditorFactory URI can be used
to specialize the cell editor that's produced and a
IPropertyEditorFactory can be used to specialize labels, i.e, although
the type in the model might be long, the label that's shown for it can
display a date, formatted in any way you like.

While working on the validation it became clear that many things being
validated rely on what's in the target platform but instead are
validating against what's registered the running IDE.   CommonPlugin now
provides basic support for reading information from the target platform
and provides a base registry implementation that can easily be extended
to make use of this.  EcorePlugin is changed to reuse this new basic
implementation to avoid having it's own large PDEHelper class.  All this
is of course done reflectively so no PDE need be installed, in which
case, the behavior is to return what's in the running IDE.

Several problems in the auto resizing of columns and the auto expansion
of property in the properties view are fixed.

EFactoryImpl is changed to use a registered conversion delegate first
before trying other approaches.  This makes it more efficient to use
conversion delegates and is more in keeping with the intent that such a
declaration always specialize the behavior of conversion.