[537093] Fix Ecore item providers to properly filter children features
Rather than filtering the results of getChildren, we need to filter the
results of the getChildrenFeatures.  Otherwise the factoring computation
when factoring a move command will consider children of features that
are not actually displayed, resulting in an incorrect index computation.

Also fix problems with inconsistent handling of the check action state
for Show Generics.