[543340] Add access of inverse references by a specific feature
Improve the logic for the new method.  Specifically there are
effectively only three cases, container, containment, and cross
reference, so we don't ever need to do all three approaches of gathering
settings.  Also, we should be careful that the argument feature's
opposite is truly a feature of the argument EObject.