Bug: some readObject method are synchronized
This serializable class defines a readObject() which is synchronized.
By definition, an object created by deserialization is only reachable by
one thread, and thus there is no need for readObject() to be
synchronized.  If the readObject() method itself is causing the object
to become visible to another thread, that is an example of very dubious
coding style.

The question is where the synchronize should be on. Where it is
currently might not be the right location to work as fast as possible.

Rank: Troubling (14), confidence: Normal
Pattern: RS_READOBJECT_SYNC
Type: RS, Category: MT_CORRECTNESS (Multithreaded correctness)

Change-Id: I596d5b69c6c3d97bcd28e61134c24924ea7e9a72
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>