Register EMF-packages declared in MANIFEST.MF in EPackage-registry
if they are not already registered through an Eclipse Extension in a
plugin.xml.

Add an OSGi Bundle-Tracker that registers the EPackages declared in the
MANIFEST.MF of a bundle by reading the provided capabilities in the
'org.eclipse.emf.ecore.generated_package' namespace.
EMF can already generate these capabilities as of Bug 581348 [1]

The EPackages declared in the MANIFEST are processed after the EPackages
potentially declared in a plugin.xml and existing registrations with an
equal URI are not overridden to grand registration in plugin.xml files
precedence and to ensure backwards compatibility.

If EMF is not executed in a OSGi runtime the corresponding header of all
MANIFEST.MF files discoverable by the provided classloader are read
directly to ensure EPackages are then registered too.

[1] - https://bugs.eclipse.org/581348