resolve EClass of DyanmicEObject if eClass is proxy
If a dynamically loaded EPackage is unloaded and then reloaded, the
Class becomes a Proxy. Independetly loaded DynamicEObjects will then
return detached and out of date EClasses, which can cause a lot of
issues, as normally nobody will expect this. With this, we at least make
an effort to resolve the Proxy. If the new Version of the EClass is too
different, this will cause Errors, but then you are in world of self
inflicted Pain anyway.

Signed-off-by: Juergen Albert <j.albert@data-in-motion.biz>